% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCopemanFunctions.R
\name{calcCopeman_GrowthRate}
\alias{calcCopeman_GrowthRate}
\title{Calculate temperature-dependent snow crab growth rates}
\usage{
calcCopeman_GrowthRate(T)
}
\arguments{
\item{T}{- vector of temperatures}
}
\value{
vector of growth rates
}
\description{
Function to calculate temperature-dependent snow crab growth rates.
}
