% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCopemanFunctions.R
\name{calcCopeman_Survival}
\alias{calcCopeman_Survival}
\title{Calculate temperature-dependent snow crab cumulative survival, C1-C5}
\usage{
calcCopeman_Survival(T)
}
\arguments{
\item{T}{- vector of temperatures}
}
\value{
vector of survival probablities
}
\description{
Function to calculate temperature-dependent snow crab cumulative survival, C1-C5.
}
