% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntermoltDuration_Belehradek.R
\name{calcIMDbyStage_FixedT}
\alias{calcIMDbyStage_FixedT}
\title{Calculate the IMD by stage for fixed T}
\usage{
calcIMDbyStage_FixedT(
  T,
  stages = c("Z1", "Z2", "M"),
  source = "Ouellet and Ste Marie 2018"
)
}
\arguments{
\item{T}{- vector of (fixed) temperatures}

\item{stages}{- character vector of stages to calculate IMD for}

\item{source}{- character value with source for constants}
}
\value{
a tibble with stage, temperature, and intermolt duration, in the time units of \code{a}.
}
\description{
Function to calculate the IMD by stage for fixed T.
}
\details{
Temperatures are regarded as fixed, so this function is appropriate to
use if T represents temperatures spatially, for example.
See also \code{\link{calcIntermoltDuration_Belehradek}}.
}
