% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntermoltDuration_Belehradek.R
\name{calcIntermoltDuration_Belehradek}
\alias{calcIntermoltDuration_Belehradek}
\title{Calculate the intermolt duration using the Belehradek equation}
\usage{
calcIntermoltDuration_Belehradek(a, b, c, T, dt = 1)
}
\arguments{
\item{a}{- the "thermal constant" (Ouellet and Ste. Marie, 2018)}

\item{b}{- the "threshold temperature constant" (Ouellet and Ste. Marie, 2018)}

\item{c}{- exponent for Belehradek equation (negative of Ouellet and Ste. Marie, 2018 value)}

\item{T}{- temperature (either a single number or a time series)}

\item{dt}{- time step (same time units as a), if T is a time series}
}
\value{
The intermolt duration, in the time units of \code{a}.
}
\description{
Function to calculate the intermolt duration using the Belehradek equation.
}
\details{
The intermolt duration using Belehradek equation is \eqn{D = a/(T-b)^c} for development at constant T.
Note that this => \eqn{1/D = [(T-b)^c]/a} => \eqn{S (1/D) dt = S [(T-b)^c]/a dt} => \eqn{1 = S [(T-b)^c]/a dt}, where S is the
time integral from 0 to D. For variable T, then, the intermolt duration is assumed to be
given by the time at which the integral \eqn{S [(T-b)^c]/a dt = 1}. Since a is constant, this condition is equivalent
to \eqn{S [(T-b)^c] dt = a}. Note that Reamur's Law can be obtained by setting the exponent "c" to 1.
}
