% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntermoltDuration_Belehradek.R
\name{calcTempForIMD_Belehradek}
\alias{calcTempForIMD_Belehradek}
\title{Calculate the temperature corresponding to an intermolt duration using Belehradek's function}
\usage{
calcTempForIMD_Belehradek(a, b, c, D)
}
\arguments{
\item{a}{- the "thermal constant" (Ouellet and Ste. Marie, 2018)}

\item{b}{- the "threshold temperature constant" (Ouellet and Ste. Marie, 2018)}

\item{c}{- exponent for Belehradek equation (negative of Ouellet and Ste. Marie, 2018 value)}

\item{D}{- the intermolt duration (either a single number or a time series)}
}
\value{
The intermolt duration, in the time units of \code{a}.
}
\description{
Function to calculate the temperature corresponding to an intermolt duration using Belehradek's function.
}
\details{
The intermolt duration using Belehradek's function is \eqn{D = a/(T-b)^c} for development at constant T, so
\eqn{T = (a/D)^(1/c)+b} for a given intermolt duration..
}
