% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntermoltDuration_Belehradek.R
\name{calcTforStageAndIMD_Belehradek}
\alias{calcTforStageAndIMD_Belehradek}
\title{Calculate the temperature corresponding to an intermolt duration using Belehradek's function}
\usage{
calcTforStageAndIMD_Belehradek(sources = "all", stages = "all", D)
}
\arguments{
\item{sources}{- character value with source for constants}

\item{stages}{- character vector of stages to calculate IMD for}

\item{D}{- the intermolt duration (either a single number or a time series)}
}
\value{
A tibble, with coloumns 'source', 'stage','T', and 'D'.
}
\description{
Function to calculate the temperature corresponding to an intermolt duration 
using Belehradek's function for given stages and sources.
}
\details{
The intermolt duration using Belehradek's function is \eqn{D = a/(T-b)^c} 
for development at constant T, so
\eqn{T = (a/D)^(1/c)+b} for a given intermolt duration.
}
