% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseMapLayers.R
\name{getBaseMapLayers}
\alias{getBaseMapLayers}
\title{Get \pkg{ggplot2} layers for a basemap}
\usage{
getBaseMapLayers(czType = c("2020", "2019"), xy = FALSE, bw = FALSE)
}
\arguments{
\item{czType}{- type of connectivity zone to get ("2020"or "2019")}

\item{xy}{- flag (T/F) for coordinate type (TRUE=Alaska Albers projection; FALSE=WGS84 lat/lon)}

\item{bw}{- flag (T/F) for black & white or color scheme}
}
\value{
list with ggplot2 layers 'bathym', 'land', 'zones', 'labels', 'map_scale', and 'theme'.
}
\description{
Function to get a set of \pkg{ggplot2} layers for a basemap.
}
\details{
If \code{xy} is TRUE, 'bathym' and 'map_scale' in the returned list will be NULL (currently).
Note that including 'theme' in the basemap simply removes the axis labels, which would otherwise be 'x' and 'y'.
}
\examples{
A default lat/lon basemap in color can be constructed using the following:\cr
require(ggplot2);
lst = getBaseMapLayers("2020",FALSE,FALSE);
ggplot()+lst$bathym+lst$land+lst$zones+lst$labels+lst$map_scale+lst$theme;

}
