% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBBox.R
\name{getBBox}
\alias{getBBox}
\title{Get the standard bbox for maps of model output}
\usage{
getBBox()
}
\value{
\code{sf}-style version of standard bounding box in WGS84
}
\description{
Function to get the standard bbox for maps of model output
}
