% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntermoltDuration_Belehradek.R
\name{getBelehradekConstants}
\alias{getBelehradekConstants}
\title{Get the constants for the Belehradek equation by life stage and source}
\usage{
getBelehradekConstants(sources = "all", stages = "all")
}
\arguments{
\item{sources}{- "all", "Ouellet and Ste Marie 2018", or "Yamamoto et al. 2014"}

\item{stages}{- "all" or character vector with required life stages}
}
\value{
tibble with columns 'source', 'stage', 'a', 'se_a', 'b', 'se_b', and 'c'.
}
\description{
Function to get the constants for the Belehradek equation by life stage and source.
}
\details{
Note that Reamur's equation is equivalent to the Belehradek equation with
the constant \eqn{c = 1}.
}
