% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConnectivityZones.R
\name{getConnectivityZones}
\alias{getConnectivityZones}
\title{Get a set of connectivity zones}
\usage{
getConnectivityZones(
  type = c("2020", "30-layer", "2019", "10-layer"),
  crs_type = c("LL", "AA")
)
}
\arguments{
\item{type}{- type of connectivity zone to get ("2020","30-layer","2019", or "10-layer")}

\item{crs_type}{- type of crs ("LL" for lat/lon, "AA" for Alaska Albers projection)}
}
\value{
\pkg{sf} dataframe with connectivity zones in given crs.
}
\description{
Function to get a set of connectivity zones as a \pkg{sf} dataframe.
}
\details{
None.
}
