% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConnectivityZonesWeb.R
\name{getConnectivityZonesWeb}
\alias{getConnectivityZonesWeb}
\title{Get a \pkg{sf} dataframe representing a directed web for a set of connectivity zones}
\usage{
getConnectivityZonesWeb(
  type = c("2020", "30-layer", "2019", "10-layer"),
  crs_type = c("LL", "AA")
)
}
\arguments{
\item{type}{- type of connectivity zone to get ("2020","30-layer","2019", or "10-layer")}

\item{crs_type}{- type of crs ("LL" for lat/lon, "AA" for Alaska Albers projection)}
}
\value{
\pkg{sf} dataframe with a directed web for the connectivity zones in given crs.
}
\description{
Function to get a \pkg{sf} dataframe representing a directed web for a set of connectivity zones.
}
\details{
None.
}
