% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLifeStageInfo.R
\name{getLifeStageInfo}
\alias{getLifeStageInfo}
\title{Get a list that defines life stage info for the DisMELS snow crab IBM.}
\usage{
getLifeStageInfo(resType = "NEW2.0SC")
}
\arguments{
\item{resType}{- results file type (i.e., 'NEW2.0SC')}
}
\value{
a list (see @link{@details})
}
\description{
Function to get a list that defines life stage info for the DisMELS snow crab IBM.
}
\details{
The returned list has the following elelments:
resType 
classInfo 
lifeStageType
}
