% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageData.R
\name{getPackageData}
\alias{getPackageData}
\title{Get a package layer object}
\usage{
getPackageData(name, xyType = c("LL", "AA"), bwType = c("BW", "RGB"))
}
\arguments{
\item{name}{- name of type of object to return}

\item{xyType}{- "LL" or "AA" for lat/lon or Alaska Albers coordinates}

\item{bwType}{- "BW" or "RGB" for black-and-white or color}
}
\value{
object
}
\description{
Function to get a package layer dataset
}
\details{
Following names are understood:
\itemize{
\item{Land}
\item{BathymContours}
\item{BathymLegend}
}
}
