% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntermoltDuration_Belehradek.R
\name{getTforIMD_Belehradek}
\alias{getTforIMD_Belehradek}
\title{Calculate and plot the necessary T against IMD by stage for a range of IMDs}
\usage{
getTforIMD_Belehradek(
  Ds,
  sources = "all",
  stages = c("Z1", "Z2", "M"),
  lims = NULL,
  showPlot = FALSE
)
}
\arguments{
\item{Ds}{- vector of IMDs}

\item{sources}{- character vector with sources for constants}

\item{stages}{- character vector of stages to calculate IMD for}

\item{lims}{- list with elemets 'x' and 'y' with axis limits}

\item{showPlot}{- flag (T/F) to immediately show the plot}
}
\value{
list, with a tibble (stage, temperature, and intermolt duration, in the time units of \code{a})
and a \pkg{ggplot2} object.
}
\description{
Function to calculate and plot the necessary T against IMD by stage for a range of IMDs.
}
\details{
Temperatures are regarded as fixed to determine the T corresponding to a
given IMD. See also \code{\link{calcTforStageAndIMD_Belehradek}}.
}
