% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBasemapLayers.R
\name{plotBasemapLayers}
\alias{plotBasemapLayers}
\title{Plot \pkg{ggplot2} layers of a basemap}
\usage{
plotBasemapLayers(bmls)
}
\arguments{
\item{bmls}{- basemap layers from call to \code{\link{getBasemapLayers}}}
}
\value{
list with ggplot2 plot object.
}
\description{
Function to plot a set of \pkg{ggplot2} layers for a basemap from call to \code{\link{getBasemapLayers}}.
}
\details{
None.
}
\examples{
#A default lat/lon basemap in color with bathymetry can be constructed and printed using the following:\cr
bmls = getBaseMapLayers(czType="2019",xy=FALSE,bw=FALSE,noBathym=FALSE);
print(plotBasemapLayers(bmls));

#A default Alaska Albers basemap in color can be constructed and printed using the following:\cr
#--In this case, the bathymetry (an annotation layer in lat/lon) will not be included.
bmls = getBaseMapLayers(czType="2019",xy=TRUE,bw=FALSE,noBathym=FALSE);
print(plotBasemapLayers(bmls));

}
