% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntermoltDuration_Belehradek.R
\name{plotIMDbyT_Belehradek}
\alias{plotIMDbyT_Belehradek}
\title{Calculate and plot the IMD against T by stage for a range of fixed Ts}
\usage{
plotIMDbyT_Belehradek(
  Ts,
  sources = "all",
  stages = c("Z1", "Z2", "M"),
  lims = NULL,
  showPlot = FALSE
)
}
\arguments{
\item{Ts}{- vector of (fixed) temperatures at which t}

\item{sources}{- character vector with sources for constants}

\item{stages}{- character vector of stages to calculate IMD for}

\item{lims}{- list with elements 'x' and 'y' with axis limits}

\item{showPlot}{- flag (T/F) to immediately show the plot}
}
\value{
list, with a tibble (stage, temperature, and intermolt duration, in the time units of \code{a})
and a \pkg{ggplot2} object.
}
\description{
Function to calculate and plot the IMD against T by stage for a range of fixed Ts.
}
\details{
Temperatures are regarded as fixed, so this function is appropriate to
use if T represents temperatures spatially, for example.
See also \code{\link{calcIMDforStage_Belehradekk}}.
}
