% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialIMD_calc.R
\name{spatialIMD_calc}
\alias{spatialIMD_calc}
\title{Calculate the spatial distribution of IMD for a life stage}
\usage{
spatialIMD_calc(stage, source, sf_temp, sf_crop = NULL)
}
\arguments{
\item{sf_temp}{- sigma-level temperature layer from ROMS model output as \pkg{sf} dataframe}

\item{sf_crop}{- \pkg{sf} dataframe with polygon to crop by using \code{linl[sf]{st_intersection}}}

\item{fn}{- path to ROMS model output file}
}
\value{
\pkg{sf} dataframe with spatial distribution of IMD
}
\description{
Calculate the spatial distribution of IMD for a life stage based on
temperature from ROMS model output
}
