% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialIMD_map.R
\name{spatialIMD_map}
\alias{spatialIMD_map}
\title{Map the spatial distribution of IMD for a life stage}
\usage{
spatialIMD_map(
  sf_imd,
  subtitle = "",
  scale_type_imd = "b",
  breaks_imd = ggplot2::waiver(),
  limits_imd = NULL,
  scale_type_temp = "c",
  breaks_temp = ggplot2::waiver(),
  limits_temp = NULL,
  bbx = wtsGIS::getStandardBBox("EBS"),
  land = wtsGIS::getPackagedLayer("Alaska"),
  bath = wtsGIS::getPackagedLayer("ShelfBathymetry"),
  doHist = TRUE,
  doTemp = TRUE,
  horiz = TRUE
)
}
\arguments{
\item{sf_imd}{- \pkg{sf} dataframe with IMD info (i.e., from [spatialIMD_calc])}

\item{subtitle}{- subtitle for map}

\item{scale_type_imd}{- fill scale type ("d","c","b") for IMD (default is "b"--binned)}

\item{breaks_imd}{- vector of values to use as breaks in scale for IMDs}

\item{limits_imd}{- limits to use when plotting IMD values}

\item{scale_type_temp}{- fill scale type ("d","c","b") for temperatures (default is "c"--continuous)}

\item{breaks_temp}{- vector of values to use as breaks in scale for temperatures}

\item{limits_temp}{- limits to use when plotting temperature values}

\item{bbx}{- \pkg{sf}-style bounding box}

\item{doHist}{- include histogram of IMD values as inset on IMD map}

\item{doTemp}{- add map of temperatures}

\item{horiz}{- flag (TRUE/FALSE) to arrange temperature, IMD maps horizontally}
}
\value{
ggplot2 map object or gg list object
}
\description{
Map the spatial distribution of IMD for a life stage based on
temperature from ROMS model output
}
\details{
If an input limit (limit_imd, limit_temp) is NULL, the resulting scale limit 
is: 1) based on the range of the data if the corresponding breaks is set to ggplot2::waiver() or 
2) equal to c(min(breaks),max(breaks)) if breaks is a numeric vector.
}
